pyproject.toml
[project]
name = "reversl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fa-filter>=0.2.1",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "sqlalchemy[mypy]>=2.0.41",
    "uvicorn>=0.34.2",
]

[tool.ruff]
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "data",
]
line-length = 88
indent-width = 4
fix = true

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
    "PL", # pylint
]
ignore = [
    "E501", # line too long
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
follow_untyped_imports = true
disallow_untyped_calls = false
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "data",
    "tests",
]
plugins = [
    "pydantic.mypy", 
    "sqlalchemy.ext.mypy.plugin",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
]
